Python homework 1
An attempt at python. I’ve figure out how to use if statements and “”
to search through (i.title) to look only for a specific word, and i’ve
moved in the emailcount function in, but i seem to have trouble to get
the email me part going. I made a function called emailme and i keep
ending up with errors when trying to run the code, it has something to
do with my variables not being all strings.